pyproject.toml
[project]
name = "esg-rag-agent"
version = "0.1.0"
description = "An ESG analysis system based on RAG and agentic workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.5.0",
  "joblib>=1.4.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "pypdf>=4.3.0",
  "python-docx>=1.1.0",
  "jinja2>=3.1.4",
  "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]
vectordb = [
  "chromadb>=0.5.5",
  "pymilvus>=2.5.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
esg_rag = ["web/index.html", "web/static/*.css", "web/static/*.js"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
